clean unified.test
